Rename scheduler to supervisor
